feat(questionbank): create logic filter for question

+++ backend/src/app/model/questions.py
@@ -2,7 +2,7 @@ import uuid
 from datetime import datetime
 
 from backend.src.app.core.database import Base
-from sqlalchemy import DateTime, Integer, String, func
+from sqlalchemy import DateTime, Float, Integer, String, func
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -21,7 +21,7 @@ class Questions(Base):
     options: Mapped[dict | None] = mapped_column(JSONB, nullable=False)
     correct_answer: Mapped[str] = mapped_column(String, nullable=False)
     explanation: Mapped[str | None] = mapped_column(String, nullable=True)
-    score_weight: Mapped[int] = mapped_column(Integer, nullable=False)
+    score_weight: Mapped[float] = mapped_column(Float, nullable=False)
     created_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), server_default=func.now()
     )
